created user-groups.ts file, removed dotenv from createClient
