+++ app/user_groups/page.tsx
@@ -1,35 +1,21 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import supabase from "@/actions/supabase/client";
+import { fetchUserGroups } from "@/api/supabase/queries/user-groups";
 import { UserGroup } from "@/types/schema";
 
 const UserGroupsPage = () => {
   const [userGroups, setUserGroups] = useState<UserGroup[]>([]);
 
   useEffect(() => {
-    async function fetchUserGroups() {
-      try {
-        // Pull data
-        const { data, error } = await supabase.from("user_group").select("*");
-
-        if (error) throw error;
-
-        // Process data
-        if (data) {
-          const mapped = data.map(usergrp => ({
-            id: usergrp.user_group_id,
-            user_group_name: usergrp.user_group_name,
-          }));
-
-          setUserGroups(mapped);
-        }
-      } catch (error) {
-        console.log("Error fetching orgs data from supabase API: ", error);
+    async function loadUserGroups() {
+      const fetchedUserGrps = await fetchUserGroups();
+      if (fetchedUserGrps) {
+        setUserGroups(fetchedUserGrps);
       }
     }
 
-    fetchUserGroups();
+    loadUserGroups();
   }, []);
 
   return (
